+++ incapsula/resource_policy_asset_association.go
@@ -37,6 +37,14 @@ func resourcePolicyAssetAssociation() *schema.Resource {
 				Required:    true,
 				ForceNew:    true,
 			},
+			// Optional Arguments
+			"account_id": {
+				Description: "The Asset's Account ID",
+				Type:        schema.TypeInt,
+				Computed:    true,
+				Optional:    true,
+				ForceNew:    true,
+			},
 		},
 	}
 }
@@ -47,8 +55,9 @@ func resourcePolicyAssetAssociationCreate(d *schema.ResourceData, m interface{})
 	policyID := d.Get("policy_id").(string)
 	assetID := d.Get("asset_id").(string)
 	assetType := d.Get("asset_type").(string)
+	currentAccountId := d.Get("account_id").(int)
 
-	err := client.AddPolicyAssetAssociation(policyID, assetID, assetType)
+	err := client.AddPolicyAssetAssociation(policyID, assetID, assetType, &currentAccountId)
 
 	if err != nil {
 		log.Printf("[ERROR] Could not create Incapsula policy asset association: policy ID (%s) - asset ID (%s) - asset type (%s) - %s\n", policyID, assetID, assetType, err)
@@ -60,7 +69,7 @@ func resourcePolicyAssetAssociationCreate(d *schema.ResourceData, m interface{})
 	d.SetId(syntheticID)
 	log.Printf("[INFO] Created Incapsula policy asset association with ID: %s - policy ID (%s) - asset ID (%s) - asset type (%s)\n", syntheticID, policyID, assetID, assetType)
 
-	return nil
+	return resourcePolicyAssetAssociationRead(d, m)
 }
 
 func resourcePolicyAssetAssociationRead(d *schema.ResourceData, m interface{}) error {
@@ -69,24 +78,34 @@ func resourcePolicyAssetAssociationRead(d *schema.ResourceData, m interface{}) e
 	policyID := strings.Split(d.Id(), "/")[0]
 	assetID := strings.Split(d.Id(), "/")[1]
 	assetType := strings.Split(d.Id(), "/")[2]
-
-	log.Printf("[INFO] Trying to read Incapsula Policy Asset Association: %s-%s-%s\n", policyID, assetID, assetType)
-	var isAssociated, err = client.isPolicyAssetAssociated(policyID, assetID, assetType)
+	currentAccountId := getCurrentAccountId(d, client.accountStatus)
+	if currentAccountId != nil {
+		log.Printf("[INFO] Trying to read Incapsula Policy Asset Association: %s-%s-%s for account %d\n", policyID, assetID, assetType, *currentAccountId)
+	} else {
+		log.Printf("[INFO] Trying to read Incapsula Policy Asset Association: %s-%s-%s\n", policyID, assetID, assetType)
+	}
+	var isAssociated, err = client.isPolicyAssetAssociated(policyID, assetID, assetType, currentAccountId)
 
 	if err != nil {
 		log.Printf("[ERROR] Could not read Incapsula Policy Asset Association: %s-%s-%s, err: %s\n", policyID, assetID, assetType, err)
 		return err
 	}
 
-	if isAssociated {
-		log.Printf("[INFO] Successfully read Policy Asset Association exist: %s-%s-%s\n", policyID, assetID, assetType)
-		syntheticID := fmt.Sprintf("%s/%s/%s", policyID, assetID, assetType)
-
-		d.Set("asset_id", assetID)
-		d.Set("asset_type", assetType)
-		d.Set("policy_id", policyID)
-		d.SetId(syntheticID)
+	if !isAssociated {
+		log.Printf("[ERROR] Could not find Incapsula Policy Asset Association: %s-%s-%s\n", policyID, assetID, assetType)
+		return fmt.Errorf("Incapsula Policy Asset Association doesn't exisits: policy id:%s asset id:%s asset type:%s\n", policyID, assetID, assetType)
 	}
+
+	log.Printf("[INFO] Successfully read Policy Asset Association exist: %s-%s-%s\n", policyID, assetID, assetType)
+	syntheticID := fmt.Sprintf("%s/%s/%s", policyID, assetID, assetType)
+
+	d.Set("asset_id", assetID)
+	d.Set("asset_type", assetType)
+	d.Set("policy_id", policyID)
+	if currentAccountId != nil {
+		d.Set("account_id", *currentAccountId)
+	}
+	d.SetId(syntheticID)
 
 	return nil
 }
@@ -97,8 +116,13 @@ func resourcePolicyAssetAssociationDelete(d *schema.ResourceData, m interface{})
 	policyID := d.Get("policy_id").(string)
 	assetID := d.Get("asset_id").(string)
 	assetType := d.Get("asset_type").(string)
-
-	err := client.DeletePolicyAssetAssociation(policyID, assetID, assetType)
+	currentAccountId := getCurrentAccountId(d, client.accountStatus)
+	if currentAccountId != nil {
+		log.Printf("[INFO] Trying to delete Incapsula Policy Asset Association: %s-%s-%s for account %d\n", policyID, assetID, assetType, *currentAccountId)
+	} else {
+		log.Printf("[INFO] Trying to delete Incapsula Policy Asset Association: %s-%s-%s\n", policyID, assetID, assetType)
+	}
+	err := client.DeletePolicyAssetAssociation(policyID, assetID, assetType, currentAccountId)
 
 	if err != nil {
 		return err